モバイル画面の追加ボタン表示制御
