+++ src/libs/dateFunc.ts
@@ -28,3 +28,17 @@ export const getTommorow = () => {
   format_str = format_str.replace(/DD/g, day_str);
   return format_str;
 };
+
+// 現在の日付
+export const getStringFromDate = (date: Date) => {
+  const year_str: string = date.getFullYear().toString();
+  //月だけ+1すること
+  const month_str: string = 1 + date.getMonth().toString();
+  const day_str: string = date.getDate().toString();
+
+  let format_str = "YYYY-MM-DD";
+  format_str = format_str.replace(/YYYY/g, year_str);
+  format_str = format_str.replace(/MM/g, month_str);
+  format_str = format_str.replace(/DD/g, day_str);
+  return format_str;
+};
